Make the monitor interval customizable
By default the 100 ms interval in the plugin system  will cause high CPU impact at Linux system for this clipboard plugin. With this PR the interval will be customizable for the user, and the default interval has been increased to 3000 ms for Linux system, which reduced the CPU usage in this case, and for the other system the default interval has been increased to 1000 ms, however, more frequent clipboard register will be missed.

Fixes https://github.com/tinytacoteam/zazu/issues/189

Signed-off-by: Tao Wang <twang2218@gmail.com>